Updates to config file - release 2 in progress

+++ WorkLog/Record.cs
@@ -16,7 +16,7 @@ namespace WorkLog
         public DateTime StartTime { get; set; }
         public DateTime EndTime { get; set; }
         public double TotalHours { get; set; }
-        public int ReimburseAmount { get; set; }
+        public double ReimburseAmount { get; set; }
         public string Description { get; set; }
         public string Comments { get; set; }
         public string FullRecord => $"{Client}" + "|" + $"{ProService}" + "|" + $"{Task}" + "|" + $"{Item}" + "|" + $"{Date}" + "|" + $"{StartTime}" + "|" + $"{EndTime}" + "|" + $"{TotalHours}";
